# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_write_down_what_was_said_chat_transcribe_post": {
        "properties": {
          "audio": {
            "contentMediaType": "application/octet-stream",
            "title": "Audio",
            "type": "string"
          }
        },
        "required": [
          "audio"
        ],
        "title": "Body_write_down_what_was_said_chat_transcribe_post",
        "type": "object"
      },
...
      "ChatMessageView": {
        "additionalProperties": false,
        "description": "One turn of a conversation, as the outside may read it.",
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "message_id": {
            "title": "Message Id",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "message_id",
          "role",
          "content",
          "timestamp"
        ],
        "title": "ChatMessageView",
        "type": "object"
      },
      "ChatThreadView": {
        "additionalProperties": false,
        "description": "One day's conversation, as the person would read it back.",
        "properties": {
          "editable": {
            "title": "Editable",
            "type": "boolean"
          },
          "event_date": {
            "format": "date",
            "title": "Event Date",
            "type": "string"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessageView"
            },
            "title": "Messages",
            "type": "array"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          }
        },
        "required": [
          "session_id",
          "event_date",
          "editable"
        ],
        "title": "ChatThreadView",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReviseRequest": {
        "additionalProperties": false,
        "description": "A request to say one of the earlier turns differently.",
        "properties": {
          "content": {
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "session_id": {
            "minLength": 1,
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "content"
        ],
        "title": "ReviseRequest",
        "type": "object"
      },
...
      "TranscriptView": {
        "additionalProperties": false,
        "description": "What a recording turned out to say.",
        "properties": {
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "TranscriptView",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Read access to the knowledge graph and the history of the runs that built it, and one way in: uploading an exported conversation.",
    "title": "Lumen",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/chat/days/{on}": {
      "get": {
        "description": "One earlier day, as the person would read it.\n\nBranches they moved away from are left out — this is the conversation\nthey settled on, which is also the one the pipeline extracted.",
        "operationId": "read_one_day_back_chat_days__on__get",
        "parameters": [
          {
            "in": "path",
            "name": "on",
            "required": true,
            "schema": {
              "format": "date",
              "title": "On",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatThreadView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read One Day Back",
        "tags": [
          "chat"
        ]
      }
    },
    "/chat/messages/{message_id}/revise": {
      "post": {
        "description": "Rewrite something, while the day is still open.\n\nThe rewrite is stored beside the original rather than over it, and the\nconversation starts reading from the new one. Nothing anybody said is\ndestroyed.\n\nOnce a day has been processed it refuses, and says what to do instead.\nAn episode is stored under the day it happened on rather than under what\nit says, so a re-run of an edited day would find it already saved and\nskip it — the conversation and the graph would disagree from then on with\nnothing anywhere reporting it.",
        "operationId": "say_it_differently_chat_messages__message_id__revise_post",
        "parameters": [
          {
            "in": "path",
            "name": "message_id",
            "required": true,
            "schema": {
              "title": "Message Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMessageView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Say It Differently",
        "tags": [
          "chat"
        ]
      }
    },
    "/chat/transcribe": {
      "post": {
        "description": "Turn a recording into words.\n\nA separate request rather than bytes on the socket. Audio is a bulk\nupload and the socket is for the back-and-forth; a failed upload should\nnot take a conversation down with it. What comes back is sent on as an\nordinary turn, marked as spoken.",
        "operationId": "write_down_what_was_said_chat_transcribe_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_write_down_what_was_said_chat_transcribe_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Write Down What Was Said",
        "tags": [
          "chat"
        ]
      }
    },
...
  },
...
}